Move tests to fixtures - infinity
