+++ packages/babel-plugin-minify-infinity/__tests__/fixtures/assignment-expr/actual.js
@@ -1,0 +1,1 @@
+Infinity = 1;
