+++ packages/babel-plugin-minify-infinity/__tests__/fixtures/destructure/actual.js
@@ -1,0 +1,3 @@
+({ Infinity } = 1);
+[Infinity] = foo;
+[...Infinity] = foo;
