+++ packages/babel-plugin-minify-infinity/__tests__/fixtures/fn-param/actual.js
@@ -1,0 +1,3 @@
+function a(Infinity) {}
+function b(...Infinity) {}
+function c({ Infinity }) {}
