+++ packages/babel-plugin-minify-infinity/__tests__/fixtures/property/expected.js
@@ -1,0 +1,2 @@
+var x = { Infinity: 0 };
+x.Infinity;
