+++ packages/babel-plugin-minify-infinity/__tests__/fixtures/to-1-over-0/actual.js
@@ -1,0 +1,1 @@
+Infinity;
