+++ packages/babel-plugin-minify-infinity/__tests__/fixtures/to-1-over-0/expected.js
@@ -1,0 +1,1 @@
+1 / 0;
