+++ packages/babel-plugin-minify-infinity/__tests__/index.js
@@ -1,0 +1,2 @@
+const runner = require("test-runner");
+runner(__dirname);
